made parsing of bvec and bval files more robust

+++ src/data/datasets/datasetdwi.cpp
@@ -8,7 +8,7 @@
 
 #include "datasetdwi.h"
 
-DatasetDWI::DatasetDWI( QString filename, QVector<ColumnVector>* data, QVector<float> b0Data, QVector<int>bvals, QVector<QVector3D>bvecs, nifti_image* header ) :
+DatasetDWI::DatasetDWI( QString filename, QVector<ColumnVector>* data, QVector<float> b0Data, QVector<float>bvals, QVector<QVector3D>bvecs, nifti_image* header ) :
         DatasetNifti( filename, FNDT_NIFTI_DWI, header ),
         m_data( data ),
         m_b0Data( b0Data ),
@@ -42,7 +42,7 @@ QVector<float> DatasetDWI::getB0Data()
     return m_b0Data;
 }
 
-QVector<int> DatasetDWI::getBvals()
+QVector<float> DatasetDWI::getBvals()
 {
     return m_bvals;
 }
